week09/requirements.txt
pytest>=8.0.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
-e .
